Improved underwater terrain visuals + perf, added wave rendering and shore geo to lake view

+++ engine/globals.h
@@ -8,7 +8,7 @@ const struct Vector3 GLOBAL_ORIGIN = { 0,0,0 };
 const struct Vector3 GLOBAL_UNIT = { 0,0,-1 };
 
 PlaydateAPI* pd = NULL;
-float DELTA_TIME = 0;
+double DELTA_TIME = 0;
 float TIME_SCALE = 1.f;
 
 #endif // !GLOBALS
